It's not perfect, but seems to work

+++ aura/core/source/nucleus.cpp
@@ -15,7 +15,9 @@
 // Standard includes.
 #include <chrono>
 #include <cstdint>
+#include <exception>
 #include <iostream>
+#include <limits>
 #include <mutex>
 #include <string>
 #include <thread>
@@ -37,34 +39,55 @@ namespace Aura::Core
 		ThreadPool(std::thread::hardware_concurrency()),
 		app_info({ app_name, app_major, app_minor, app_patch }),
 		ui(*this), environment(*this), render(*this),
-		frame_counter(0U), frame_limit(0U), stop(true)
+		frame_counter(0U), frame_limit(0U), rendering(false)
 	{
 		loadDisplaySettings();
-		ui.updateWindow();
-		render.createSurface();
-		render.checkPhysicalDevices();
-		render.selectPhysicalDevice();
-		render.createDevice();
-		render.createFramework();
-		render.setUpDispatch();
-		renderStart();
+		setUp(true, true);
 	}
 	/// <summary>
 	/// Stops rendering and tears-down the core.
 	/// </summary>
 	Nucleus::~Nucleus() noexcept
 	{
-		renderStop();
 		render.waitIdle();
-		{
-			std::unique_lock<std::mutex> lock { stop_guard };
-			stop_event.wait(lock, [&] { return !rendering; });
-		}
-		render.tearDownDispatch();
-		render.destroyFramework();
-		render.destroyDevice();
-		render.destroySurface();
+		tearDown(true, true);
 	}
+	/// <summary>
+	/// Restarts the renderer.
+	/// </summary>
+	void Nucleus::setUp(bool const window_reset, bool const device_reset)
+	{
+		if(window_reset)
+		{
+			ui.updateWindow();
+			render.createSurface();
+			render.checkPhysicalDevices();
+		}
+		if(device_reset || window_reset)
+		{
+			render.selectPhysicalDevice();
+			render.createDevice();
+			render.createFramework();
+		}
+		render.setUpDispatch();
+	}
+	/// <summary>
+	/// Restarts the renderer.
+	/// </summary>
+	void Nucleus::tearDown(bool const window_reset, bool const device_reset)
+	{
+		render.tearDownDispatch();
+		if(device_reset || window_reset)
+		{
+			render.destroyFramework();
+			render.destroyDevice();
+		}
+		if(window_reset)
+		{
+			render.destroySurface();
+		}
+	}
+
 	// ------------------------------------------------------------------ //
 	// Program control.
 	// ------------------------------------------------------------------ //
@@ -76,52 +99,32 @@ namespace Aura::Core
 	/// </summary>
 	void Nucleus::run(std::uint32_t const max_frames)
 	{
-		if(stop) { renderStart(); }
+		if(ui.shouldWindowClose())
+		{
+			render.waitIdle();
+			tearDown(true, true);
+			setUp(true, true);
+		}
+		std::future<void> render_task {};
 		frameCounterReset(max_frames);
 		while(!ui.shouldWindowClose())
 		{
 			ui.poolEvents();
-
-			if(renderShouldStop()) { break; }
-		}
-	}
-	/// <summary>
-	/// Frame render cycle. Queries for changes in environment, and updates
-	/// the GPU is there are any changes. Then calls renderFrame to render
-	/// and display a new frame.
-	/// </summary>
-	void Nucleus::renderCycle()
-	{
-		{
-			std::unique_lock<std::mutex> lock { stop_guard };
-			rendering = true;
-		}
-		while(true)
-		{
-			// Time stamps.
-			std::chrono::time_point<std::chrono::system_clock> start, end;
-
-			if(renderShouldStop()) { break; }
-			// If should marker time stamp.
-			if constexpr(debug_settings.frame_time)
+			if(!isRendering())
 			{
-				start = std::chrono::system_clock::now();
+				{
+					std::unique_lock<std::mutex> lock { rendering_guard };
+					rendering = true;
+				}
+				if constexpr(debug_settings.frame_time)
+				{ render_task = enqueue([&] { renderWithTime(); }); }
+				else
+				{ render_task = enqueue([&] { renderFrame(); }); }
 			}
-			bool result { render.renderFrame() };
-			if constexpr(debug_settings.frame_time)
-			{
-				end = std::chrono::system_clock::now();
-				std::cout << std::chrono::duration<double, std::milli>(end - start).count() << std::endl;
-			}
-			if(frame_limit == 0U) { continue; }
-			if(result) { frameCounterIncrement(); }
-			if(frameCounterCheck()) { renderStop(); }
+			if(frameCounterCheck()) { break; }
 		}
-		{
-			std::unique_lock<std::mutex> lock { stop_guard };
-			rendering = false;
-			stop_event.notify_all();
-		}
+		if(isRendering())
+		{ render_task.wait(); }
 	}
 	/// <summary>
 	/// Resets the frame counter and updates the limit.
@@ -137,53 +140,78 @@ namespace Aura::Core
 	/// </summary>
 	bool Nucleus::frameCounterCheck() noexcept
 	{
-		std::unique_lock<std::mutex> lock { frame_guard };
-		return frame_limit <= frame_counter;
+		if(frame_limit != 0U)
+		{
+			std::unique_lock<std::mutex> lock { frame_guard };
+			return frame_limit <= frame_counter;
+		}
+		return false;
 	}
 	/// <summary>
 	/// Increments the frame counter.
 	/// </summary>
 	void Nucleus::frameCounterIncrement() noexcept
 	{
-		std::unique_lock<std::mutex> lock { frame_guard };
-		++frame_counter;
-	}
-	/// <summary>
-	/// Render thread start.
-	/// </summary>
-	void Nucleus::renderStart()
-	{
+		if(frame_limit != 0U)
 		{
-			std::unique_lock<std::mutex> lock { stop_guard };
-			stop = false;
+			std::unique_lock<std::mutex> lock { frame_guard };
+			++frame_counter;
 		}
-		auto f_void { enqueue([&] { renderCycle(); }) };
 	}
 	/// <summary>
-	/// Checks if render should stop.
+	/// Checks rendering flag.
 	/// </summary>
-	bool Nucleus::renderShouldStop() noexcept
+	bool Nucleus::isRendering()
 	{
-		std::unique_lock<std::mutex> lock { stop_guard };
-		return stop;
+		std::unique_lock<std::mutex> lock { rendering_guard };
+		return rendering;
 	}
 	/// <summary>
-	/// Render thread stop.
+	/// Schedule a rendering job and set rendering flag to true.
 	/// </summary>
-	void Nucleus::renderStop() noexcept
+	void Nucleus::renderFrame()
 	{
-		std::unique_lock<std::mutex> lock { stop_guard };
-		stop = true;
+		if(!render.dispatchFrame())
+		{
+			std::unique_lock<std::mutex> lock { rendering_guard };
+			rendering = false;
+		}
+		if(!render.waitForMainFence(std::numeric_limits<std::uint64_t>::max()))
+		{
+			std::unique_lock<std::mutex> render_lock { rendering_guard };
+			rendering = false;
+			std::unique_lock<std::mutex> frame_lock { frame_guard };
+			frame_counter = frame_limit;
+			throw std::exception("Render timeout.");
+		}
+		else
+		{
+			std::unique_lock<std::mutex> lock { rendering_guard };
+			rendering = false;
+		}
+		frameCounterIncrement();
 	}
+	/// <summary>
+	/// Schedule a rendering job, wait for it to stop and outputs time.
+	/// </summary>
+	void Nucleus::renderWithTime()
+	{
+		std::chrono::time_point<std::chrono::system_clock> start, end;
+		start = std::chrono::system_clock::now();
+		renderFrame();
+		end = std::chrono::system_clock::now();
+		std::cout << std::chrono::duration<double, std::milli>(end - start).count() << std::endl;
+	}
+
 	// ------------------------------------------------------------------ //
 	// Nucleus update calls.
 	// ------------------------------------------------------------------ //
 	/// <summary>
 	/// Updates the debug settings.
 	/// </summary>
-	void Nucleus::updateDebugSettings(DebugSettings const & new_settings)
+	void Nucleus::updateDebugSettings(DebugSettings new_settings)
 	{
-		debug_settings = new_settings;
+		debug_settings = std::move(new_settings);
 	}
 	/// <summary>
 	/// Updates the display settings. 
@@ -191,7 +219,7 @@ namespace Aura::Core
 	/// it's dimensions requires an window reset and a render re-build.
 	/// Both changes might take some time.
 	/// </summary>
-	void Nucleus::updateDisplaySettings(DisplaySettings const & new_settings)
+	void Nucleus::updateDisplaySettings(DisplaySettings new_settings)
 	{
 		bool window_reset, device_reset, sync_reset { false };
 		{
@@ -207,40 +235,13 @@ namespace Aura::Core
 		}
 		if(!window_reset && !device_reset && !sync_reset)
 		{
-			display_settings = new_settings;
+			display_settings = std::move(new_settings);
 			return;
 		}
-		renderStop();
 		render.waitIdle();
-		{
-			std::unique_lock<std::mutex> lock { stop_guard };
-			stop_event.wait(lock, [&] { return !rendering; });
-		}
-		render.tearDownDispatch();
-		if(device_reset || window_reset)
-		{
-			render.destroyFramework();
-			render.destroyDevice();
-		}
-		if(window_reset)
-		{
-			render.destroySurface();
-		}
-		display_settings = new_settings;
-		if(window_reset)
-		{
-			ui.updateWindow();
-			render.createSurface();
-			render.checkPhysicalDevices();
-		}
-		if(device_reset)
-		{
-			render.selectPhysicalDevice();
-			render.createDevice();
-			render.createFramework();
-		}
-		render.setUpDispatch();
-		renderStart();
+		tearDown(window_reset, device_reset);
+		display_settings = std::move(new_settings);
+		setUp(window_reset, device_reset);
 	}
 	/// <summary>
 	/// Loads saved display settings.
